pkg/reportingplugins: remove; moved to chainlink-mercury
