--- pkg/reportingplugins/README.md
@@ -1,3 +1,0 @@
-# OCR2 ReportingPlugins
-
-This directory contains OCR2 `ReportingPlugin` packages, with implementations of [ReportingPlugin](https://pkg.go.dev/github.com/smartcontractkit/libocr@v0.0.0-20230112103111-f100435787c8/offchainreporting2/types#ReportingPlugin) & [ReportingPluginFactory](https://pkg.go.dev/github.com/smartcontractkit/libocr@v0.0.0-20230112103111-f100435787c8/offchainreporting2/types#ReportingPluginFactory).
