--- pkg/reportingplugins/mercury/v2/mercury_observation_v2.proto
@@ -1,19 +1,0 @@
-syntax="proto3";
-
-package mercury_v2;
-option go_package = ".;mercury_v2";
-
-message MercuryObservationProto {
-    uint32 timestamp = 1;
-
-    bytes benchmarkPrice = 2;
-    bool pricesValid = 3;
-
-    int64 maxFinalizedTimestamp = 4;
-    bool maxFinalizedTimestampValid = 5;
-
-    bytes linkFee = 6;
-    bool linkFeeValid = 7;
-    bytes nativeFee = 8;
-    bool nativeFeeValid = 9;
-}
