+++ pkg/types/mercury/v3/types.go
@@ -1,19 +1,12 @@
-package mercury_v2 //nolint:revive
+package v3
 
 import (
 	"math/big"
 
 	ocrtypes "github.com/smartcontractkit/libocr/offchainreporting2plus/types"
 
-	"github.com/smartcontractkit/chainlink-common/pkg/reportingplugins/mercury"
+	"github.com/smartcontractkit/chainlink-common/pkg/types/mercury"
 )
-
-type PAO interface {
-	mercury.PAO
-	GetMaxFinalizedTimestamp() (int64, bool)
-	GetLinkFee() (*big.Int, bool)
-	GetNativeFee() (*big.Int, bool)
-}
 
 type ReportFields struct {
 	ValidFromTimestamp uint32
@@ -22,6 +15,8 @@ type ReportFields struct {
 	LinkFee            *big.Int
 	ExpiresAt          uint32
 	BenchmarkPrice     *big.Int
+	Bid                *big.Int
+	Ask                *big.Int
 }
 
 // ReportCodec All functions on ReportCodec should be pure and thread-safe.
@@ -38,4 +33,15 @@ type ReportCodec interface {
 	MaxReportLength(n int) (int, error)
 
 	ObservationTimestampFromReport(ocrtypes.Report) (uint32, error)
+}
+
+type Observation struct {
+	BenchmarkPrice mercury.ObsResult[*big.Int]
+	Bid            mercury.ObsResult[*big.Int]
+	Ask            mercury.ObsResult[*big.Int]
+
+	MaxFinalizedTimestamp mercury.ObsResult[int64]
+
+	LinkPrice   mercury.ObsResult[*big.Int]
+	NativePrice mercury.ObsResult[*big.Int]
 }
